Break up session into file
